refactor: separate logic and view of login page
