+++ src/index.css
@@ -655,11 +655,6 @@ video {
   margin: auto;
 }
 
-.my-4{
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-}
-
 .my-2{
   margin-top: 0.5rem;
   margin-bottom: 0.5rem;
@@ -707,6 +702,14 @@ video {
 
 .ml-1{
   margin-left: 0.25rem;
+}
+
+.ml-12{
+  margin-left: 3rem;
+}
+
+.ml-14{
+  margin-left: 3.5rem;
 }
 
 .flex{
@@ -769,10 +772,6 @@ video {
 .w-max{
   width: -webkit-max-content;
   width: max-content;
-}
-
-.w-325{
-  width: 325px;
 }
 
 .w-2\/4{
@@ -872,16 +871,6 @@ video {
   border-color: rgba(55, 65, 81, var(--tw-border-opacity));
 }
 
-.border-green-500{
-  --tw-border-opacity: 1;
-  border-color: rgba(16, 185, 129, var(--tw-border-opacity));
-}
-
-.border-red-500{
-  --tw-border-opacity: 1;
-  border-color: rgba(239, 68, 68, var(--tw-border-opacity));
-}
-
 .bg-gray-100{
   --tw-bg-opacity: 1;
   background-color: rgba(243, 244, 246, var(--tw-bg-opacity));
@@ -896,12 +885,18 @@ video {
   padding: 1rem;
 }
 
-.p-8{
-  padding: 2rem;
-}
-
 .p-2{
   padding: 0.5rem;
+}
+
+.py-4{
+  padding-top: 1rem;
+  padding-bottom: 1rem;
+}
+
+.px-2{
+  padding-left: 0.5rem;
+  padding-right: 0.5rem;
 }
 
 .px-4{
@@ -914,29 +909,14 @@ video {
   padding-bottom: 0.75rem;
 }
 
-.py-4{
-  padding-top: 1rem;
-  padding-bottom: 1rem;
+.px-8{
+  padding-left: 2rem;
+  padding-right: 2rem;
 }
 
 .py-2{
   padding-top: 0.5rem;
   padding-bottom: 0.5rem;
-}
-
-.px-2{
-  padding-left: 0.5rem;
-  padding-right: 0.5rem;
-}
-
-.px-6{
-  padding-left: 1.5rem;
-  padding-right: 1.5rem;
-}
-
-.px-8{
-  padding-left: 2rem;
-  padding-right: 2rem;
 }
 
 .pt-1{
@@ -1134,8 +1114,8 @@ video {
 }
 
 @media (min-width: 640px){
-  .sm\:ml-12{
-    margin-left: 3rem;
+  .sm\:ml-14{
+    margin-left: 3.5rem;
   }
 
   .sm\:block{
@@ -1171,12 +1151,12 @@ video {
     width: 250px;
   }
 
-  .sm\:w-375{
-    width: 375px;
-  }
-
   .sm\:w-11\/12{
     width: 91.666667%;
+  }
+
+  .sm\:w-375{
+    width: 375px;
   }
 
   .sm\:flex-row{
